requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
pre-commit==2.20.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.3
email-validator==1.3.0
pydantic==1.10.2
bcrypt==4.0.0
pyjwt==2.5.0
python-multipart==0.0.5
python-dotenv==0.10.4
